Allow configuration of var cache chunk size.
This should resolve issues with chunk size incompatibility between
autodiff Jacobian calls and the diagnostic VarCache.

Unfortunately, this change is moot because there seems to be some
weird issue with DiffEqBase.DiffCache where get_tmp is an order of
magnitude slower than just allocating a new vector at each timestep.
As such, we use the latter appraoch for now (even though it is
theoretically less efficient.