Refactor state handling and model builder
This is a complete overhaul of the internal state handling system in `LandModel`.

The main goal was to decouple discretized ("on grid") state variables from the stratigraphy.

Grid variables are now stored and handled separately, allowing for more flexibility in modifying the stratigraphy mid-simulation.

Most notably, layer boundaries are now fully treated as parameters which can be tuned without rebuilding/recompiling the model.

While dynamic grid/stratigraphies are still not directly supported, this new state system provides a better basis for providing this functionality in the future.