Add param method with global auto-param switch
Adds a new method `param` and type `FixedParam` which represents a parameter
of the system that is assumed constant at simulation time. This is useful
for tracking all parameters of the system but still permitting the user
to individually specify which values to vary as parameters.

The global switch `AUTOPARAM` can be enabled to mark all parameters
as variable by default.